Refactor ZekrCard - Add vibration feature
